+++ src/main/java/domain/board/BoardFactory.java
@@ -4,9 +4,8 @@ import static domain.Position.INITIAL_POSITION;
 import static domain.Position.X_MAXIMUM_POSITION;
 import static domain.Position.Y_MAXIMUM_POSITION;
 
-import domain.CountryType;
+import domain.country.CountryType;
 import domain.Position;
-import domain.TableSetting;
 import domain.piece.Cannon;
 import domain.piece.Chariot;
 import domain.piece.Elephant;
