+++ src/main/java/domain/country/CountryType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.country;
 
 public enum CountryType {
     HAN,
