+++ src/test/java/domain/board/TableSettingTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.board;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
